11/4/14 Added setting for allowing tile resizing to be locked to the aspect ratio of the video (takes into account objects that are outside the video).
Modifications to TileObject and drawing of tile when resizing to base it
on the aspect ratio of the video (when applicable).

Changed behaviour of video aspect ratio. Previously for the video region
to have a default aspect ratio a WSS/AFD mode needed to be enabled, and
the default aspect ratio in the WSS tab needed to be set. Now, it does
not depend on any WSS/AFD mode being enabled. The video aspect ratio
setting is now in the type tab (along with the tile aspect ratio
setting). There is a checkbox for locking the tile resizing to the video
aspect ratio. If this is ticked and the video aspect ratio is set to
anything other than "None" then the Tile Aspect Ratio setting is
ignored, and the resizing will be done according to the video aspect
ratio, taking into account objects outside the video.